git ignore .cache dir and compile_commands.json
Both of these are somewhat specific to emacs / LSP / platformio.

The cache dir contains files stored by clang

The compile commands assists the LSP in understanding source files,
specifically the framework files outside this project's directory.
It is generated using the command `pio run -t compiledb`

Neither are relevant to the distribution, and are specific to the
programmer's environment.