fix #10: handle wifi disconn prior to deep sleep
When wifi was operating in station mode (e.g. connected to a home
network), radio timeout would cause and error in the wifi-handler

This error arises because of confusion in how to handle an incoming wifi
disconnect event.  We had been counting "disconnected" events to figure
out when to swap over to access-point mode, but during deep-sleep
shutdown, we of course don't want the wifi disconnection to do that.

This commit sifts the logic a bit to handle the two senses of
"disconnected" more finely.