create partition table to expand code region
For seeed_xiao_esp32c3, and perhaps in the future other boards as well,
we increase the code partition to allow for growing firmware.

Justin K5EM had reported his recent image not fitting. We increase it
from 1M to 1.44M, (in honor of floppy disks ;)

(Presently, we still have a double-partition to still allow for OTA one
day. If need be, we can further consolidate if we restrict ourselves to
serial updates only, but that could stymie adoption by non-coders.)