fix #19: protect access to radio with mutex
Implement a global mutex, KXCommunicationMutex that can be used to
provide exclusive access to the radio, ensuring two things:
1) incoming commands from the web don’t clobber each other
2) critical ft8 timing won’t be interrupted by other trivial requests

This code is a little fragile, and requires that programmers put the
mutex around critical code.  If we move to C++, there’s an opportunity
to encapsulate more into the locking, and then assert appropriate
locking within the actual uart comm with the radio.