more sophisticated locking
Use a BasicLockable-conforming class, here called Lock, to manage the
mutex on access to the radio's communication (ACC) port.

In adition to providing lock/unlock to provide exlusive access, we also
check at the point of actually accessing that port whether the lock was
in place, providing some assurance that code is correct.  We raise a
console message if that is not the case.

Additionally, this new construct allows for raii-idiomed locking, as in:
{
    const std::lock_guard<Lock> lock(RadioPortLock);  // raii does the lock
    /* communicate with the radio safely */
} // RadioPortLock is unlocked as it falls out of scope