more sophisticated logging
TAG is now differentiated by file, which is meant to indicate the
sub-module of SOTAcat.  The new format for tags is "TAG8", meaning an
eight-letter suffix to "sc:", as in "sc:wifi....".  Right-pad the suffix
so that the subsequent colons will all line up and the logs will be
pretty.

By increasing the default log level only for our module, we can now use
the full panoply of log levels, so this commit also differentiates
clearly between verbose, info, warning, and error messages.