fix #20: add radio clock sync button to settings
New handler for setting time on the radio, based on the UTC time in the
browser.  See the "Sync Time" button in the Settings tab.

This commit creates a couple new entry points into the kx_commands.  In
doing so, I've become more aware of the need to do three things:
1) move the warnings about "failing to lock the radio port" up from the
uart-communication level to the kx-interface level.
2) refactor the kx communication code a bit. There is some magic testing
and setting in there baesd on command, command type, expected arguments
and their lengths.  While functional, it's challenging to read, debug,
or extend.
3) perhaps enclose this in a class that clarifies roles of some methods
and variables.