add distance column to sota/pota lists
Adds distance from approximate current position to the spot, in Km.
Note that this isn't meant to be precise, but does give an estimate of
how far the signal must propagate.

Sadly, since we're not serving HTTPS, we can't use HTML5 navigator
geolocation to find our address, so we'll have to use a call to a
webservice, which likely returns the location of our internet provider,
which is almost certainly further away.