consolidate spot details, prepare for sorting
Enrich spot entries with more details.  Previously, we had added
distance.  We now also add a canonical timestamp, the operator's base
callsign (minus an suffixes), and whether the spot is an older duplicate
of another more recent spot.

With those additions, initial sorting by UTC time is trivial (and in
fact comes "for free" due to the duplicate annotation), and display code
can be simplified.

All this in preparation for column sorting and selection, to come in a
subsequent commit.