fix #22: selectable refresh rate; pota de-dup
This implements a user-selectable refresh rate via dropdown menu
available on both sota and pota pages.  The two are syncronized and
saved across browser sessions (locally, in the browser, not the
SOTAcat).

While I was creating duplicated code ;), I added the de-dup checkbox to
the pota page as well.

[Soon I hope to de-dup the sota/pota javascript code.  At least the table
filling should be consolidated once the spots data structure is fully
augmented with information, and the html headers can be used to select
relevant fields for display.]