BREAKING CHANGE: Implement persistent settings.
The settings tab in the web UI now controls Wifi connection credentials.
No longer do we depend on the WIFI environment variables.
If you had relied upon those in the past, then please:
1. Connect to SOTAcat-1234, where 1234 is the mac-unique id of your hw
2. Go to the settings tab
3. Input relevant wifi credentials for your networks
4. Reboot your sotacat (power-cycle)

This method now enables starting from scratch, so that more novice users
can simply download binary firmware, install using the web-method that
Brian Mathews implemented recently, and be ready to connnect and
configure.

Note too that there is now a "Factory Reset" button on the Settings page
which simply erases the non-volatile storage we use to remember the
settings, and restarts the SOTAcat.