Fully refactored kx radio
Fix #37
Fix #29

Moved all KX code to the `class KXRadio`, which is now instantiated as a
globally-available singleton `kxRadio`.

In doing so, this eliminates the uart_connect.{cpp,h} files.

Reworked the Lock class into `Lockable`, which is now it's own generic
base class, from which KXRadio is derived.
Appropriate lock state is now checked in public methods of KXRadio we
propose as the API to the radio hardware.  This addresses #29

KXRadio now has a method `is_connected` corresponding to the successful
association between the radio and SOTACAT.  This is used in the
webserver, which will deflect webservice calls that require the radio
when it is not yet set up.  This addresses #37

The original cause of #37 was a lock that had been acquired to detect
the baud rate of the SOTACAT/KX ACC connection, blocking other
webservice calls requiring the radio, which in effect blocked the
webserver entirely and prohibited access to the settings page.

Partially implemented in this commit, and still on-going, is a
rationalization of #included files.  Some files don't need to be
included.  Others would benefit from angle brackets rather than double
quote delimeters.