Fix #22 - drop-down history duration for spot list
Removed code that had implemented variable refresh rate.  Refresh is now
every minute, as it was earlier.

Implemented a drop-down to select varying history duration for
displaying the spot list.  Two things to note:
1)  The dropdown is only for SOTA, not POTA.  POTA has no such mechanism
to request spots within a time range, so we just get what appears to be
the spots within the last hour.  My understanding is that fresher spots
are those that are respotted, and those that are more than an hour old
are stale, in practice.
2)  SOTA spot history is quantized to an hour, so this commit doesn't
implement the 15 min and 30 min options as requested in the enhancement
request.  I had experimented with a secondary filter to get
finer-grained narrowing, but given that the list is sortable by time, it
didn't seem to be worth the code complication.

In the course of implementing this, I chose to simplify the
refreshSotaPotaJson routine so that it only gets the data relevant to
the current tab.  This should lighten web traffic. Switching tabs
between SOTA and POTA always fetches the latest data anyway, so there's
no reason to keep the "other" tab's data up-to-date.