propose standard formatting for cpp/h files
By chance, we've been adhering pretty closely to a standard format in
our work so far.  But for those of use (me) with format-before-save
hooks installed in their editors, I've recently noticed more unexpected
changes further away from the source I was actually editing.

A codified coding standard helps eliminate these spurious changes.

The .clang-format that I commit here is very close to the latent
standard we've been following so far.  It is based on the "LLVM"
standard.  The deviations from that standard are described in the file.
Most notable are
1) "BraceWrapping:/ BeforeElse: true", which means "else" will always
appear on a new line.  Other than that, open braces appear on the same
line as the opening statement.
2) "SpaceBeforeParensOptions:/ BeforeNonEmptyParentheses: true", which
means there will always be a space between an identifier and a paren of
params or args.
These are my own personal preferences, matching decades of code.  Brian,
these habits track back to Ithaca Software days.  (Though I won't
subject you to "stack", and "alter" embellishments, despite my strong
desire for symmetry in the code.)
The remainder of the diffs from LLVM are alignment and spacing
improvements that are likely universally acceptable.

I have NOT run this formatting against any files and committed them yet.
That will be the next step, once others have a chance to see the effect
of this change.