close #41: clarify publication of wifi address
With Brian's discovery of the Bonjour Browser app on android, it is no
longer necessary to scroll the client address on the VFO.  It's far
easier to just use the Bonjour Browser app and click on the respective
device address.

This commit adds some links to the Bonjour Browser app for installation
convenience, and corrects a typo.