Fix mode filter
Commit 329e8030f47b5073a5258f480b9cf66c763074a9 was a bit over-zealous
in cleaning up the style sheet.  Although the different .mode-* styles
appeared to all be the same, each inited to "{}", it's not OK to
initialize them to the /same/ "{}".  With that done, they'll all share a
style, and behave as one.  Since we want to manipulate the styles
separately, in particular to toggle their visibility on and off, we must
have separate styles for each.
I've added comment to explain the intent and hopefully guard against
future optimizations.