move defns from sole header into implementation
A couple definitions were promoted to header files when they were used
only once within the corresponding implementation.  They're therefore
pretty local, or even private, and not suitable as an external
interface.  Headers are trending toward interfaces - either to the
outside world entirely as an API, or moreso in our case, a contract
between implementations, which should be as minimal as possible.