fix #42: minimal changes for K5EM hw support
Use runtime rather than compile-time determination of hardware,
and initialize hardware specific constants based on that hardware.
This allows us to keep just one unified build, but does require that we
find a way to detect the hardware type, which should be done in the
routine ```detect_hardware_type``` in hardware_specific.cpp.

NOTE that for this commit, the hardware type is always detected as
AB6D_1.  There is a FIXME comment in the code to flag the need for a
real detection method.

Within the codebase, it's now OK to examine HW_TYPE, which will be
enum class SOTAcat_HW_Type {
    unknown,
    AB6D_1,  // the original hand-built shrink-tubed module
    K5EM_1   // the first module with battery monitor and professional casing
};
The value of HW_TYPE can be used for runtime conditionals.