fix #31 by upgrading to latest platform framework
We upgrade from espressif32 6.5 to 6.6, and release the anchored version
on mdns, allowing it to float forward.

Minor adjustments throughout the code are necessary to accommodate
stronger typing, and focused declarations in system headers.

As well, our wifi attenuation of "11" (sic), was recommended to be set
to "12" instead.

A "platformio clean" followed by a full rebuild should be all that's
necessary to bring all subordinate modules to the fore, and allow for
recompilation. If not, ```python -m pip install --upgrade
idf-component-manager``` may help to upgrade the component manager (use
```python -m pip show idf-component-manager``` to see the
currently-installed version), and ```pio update``` will update the
toolchain. `