improve server responses
Standardize server responses for both failure and success.

Typical success is now reported as a status 204, without content.  This
is contrast to former implementation which returned a 200 and some
redundant json, which could confuse the client.

Failures are now more informative, including the error status code and
associated text description, which is sent by the `httpd_resp_send_err`
function, and a descriptive message in json.  Regrettably, the content
type is set to text/plain, by the `httpd_resp_send_err`, but clients
should be able to parse the json nonetheless.

This commit also alters client functions that were reacting improperly
to return status.