fix #13: implement OTA updates
OTA firmware updates are now available through an interface at the
bottom of the Settings page.  This is available only on the SEED_XIAO
platform (which includes both AB6D_1 and K5EM_1 hardware versions),
simply because that platform has a corresponding partition table
defined.  It would be straightforward to extend support, via the
addition of an appropriate partition table.

This new implementation allows for uploading a file from client computer
to the SOTAcat.  It doesn't check for file version being valid, or more
recent, or anything, really.  It is therefore up to the user to upload
an appropriate file.  There are safeguards, however, that prevent
uploading a corrupt file, and fall back to a prior version on boot if
the newly uploaded file is not bootable.