fix #47: delay to allow power setting to complete
Setting power requires a little extra time to complete, so before we
read the power value back to confirm it was properly set, we need to
accommodate a delay.  We implement this by allowing a longer timeout
when reading the value back (getting it from the radio).

Additionally, this commit changes the implementation to compact a series
of sequential strcmp tests into a search for the target string within a
space-separated lists of candidates.  This is both more efficient, and
more readable.