Change color on [in]valid keyer text length
The keyer messages on the CAT page must not exceed 24 characters in
length.  There was already a popup dialog that would arise when the user
pressed send and the message was too long.  Additionally, the field
already had validation on-the-fly, using a pattern to count characters.
Regrettably, this realtime validation gave no feedback to the user
because I had forgotten to set the invalid/valid colors.

This commit sets the colors appropriately.

Thanks @poynting (Justin K5EM) for reporting the bug.