fix #66: setting power to max achieves best max
Formerly, we arbitrarily set power to 10w when max was requested in ui.
This is because if setting to another higher value failed, we'd log an
error and return a failure to the UI.

In this commit, we request 15w as the max (matching the largest value
between KX2 and KX3), and gracefully handle the radio's fallback to a
lower value.