Don't call radio-bound apis if on localhost
This improvement targets debugging sessions served from the local
computer rather than SOTAcat hardware.  In that case, we're running a
computer, not SOTAcat, and with no possibility of contacting the radio.
We wish to avoid the failures of *trying* to contact and the resulting error
messages cluttering the console log.  So, abort those suspect calls
early to prevent certain error.