Feature: override apparent location
Background: It's a drag that we can't use the built-in GPS apis in the
browser, which require ssl.  We're not serving the page from SOTAcat via
https, so it's not possible to access that api.  As a backup, we've
been using IP-based geolocation, which can be wildly inaccurate
depending on our location relative to cell towers and their backhaul
stations.
So,
This feature allows us to override our geolocation, which can be
copy-pasted from SOTAmat or any other source.  Doing so will improve the
distance calculations in all the spots in sota/pota lists.